control-tower/requirements.txt
streamlit>=1.32.0
streamlit-autorefresh>=1.0.1
python-dotenv>=1.0.0
pandas>=2.0.0
pyarrow>=14.0.0
boto3>=1.34.0
plotly>=5.18.0
numpy>=1.24.0
matplotlib>=3.10.9
scipy>=1.17.1
deltalake>=0.20.0
scikit-learn>=1.2
joblib>=1.3
